🐛 remove download script from Dockerfile
For testing and getting the action to work with docker it is easier to
run the downloading of the readme-scraper script from within the action.
So each time the action runs it fetches the newest version of the
script. Otherwise If I have this steps inside the Dockerfile, everytime
I adjust the script I would need to build tag and push the changes back
to the registry.